+++ ikea-zigbee-drivers/Ikea_E1743.groovy
@@ -9,7 +9,7 @@ import groovy.transform.Field
 import com.hubitat.zigbee.DataType
 
 @Field static final String DRIVER_NAME = 'IKEA Tradfri On/Off Switch (E1743)'
-@Field static final String DRIVER_VERSION = '5.3.0'
+@Field static final String DRIVER_VERSION = '5.4.0'
 
 // Fields for capability.HealthCheck
 import groovy.time.TimeCategory
@@ -53,7 +53,7 @@ metadata {
             name:'helpInfo', type:'hidden',
             title:'''
             <div style="min-height:55px; background:transparent url('https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/img/Ikea_E1743.webp') no-repeat left center;background-size:auto 55px;padding-left:60px">
-                IKEA Tradfri On/Off Switch (E1743) <small>v5.3.0</small><br>
+                IKEA Tradfri On/Off Switch (E1743) <small>v5.4.0</small><br>
                 <small><div>
                 • <a href="https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/#tradfri-onoff-switch-e1743" target="_blank">device details</a><br>
                 • <a href="https://community.hubitat.com/t/release-ikea-zigbee-drivers/123853" target="_blank">community page</a><br>