+++ ikea-zigbee-drivers/Tuya_TS0201.groovy
@@ -9,7 +9,7 @@ import groovy.transform.Field
 import com.hubitat.zigbee.DataType
 
 @Field static final String DRIVER_NAME = 'Tuya Temperature and Humidity Sensor (TS0201)'
-@Field static final String DRIVER_VERSION = '5.3.0'
+@Field static final String DRIVER_VERSION = '5.4.0'
 
 // Fields for capability.HealthCheck
 import groovy.time.TimeCategory
@@ -44,7 +44,7 @@ metadata {
             name:'helpInfo', type:'hidden',
             title:'''
             <div style="min-height:55px; background:transparent url('https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/img/Tuya_TS0201.webp') no-repeat left center;background-size:auto 55px;padding-left:60px">
-                Tuya Temperature and Humidity Sensor (TS0201) <small>v5.3.0</small><br>
+                Tuya Temperature and Humidity Sensor (TS0201) <small>v5.4.0</small><br>
                 <small><div>
                 • <a href="https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/#tuya-temperature-and-humidity-sensor-ts0201" target="_blank">device details</a><br>
                 • <a href="https://community.hubitat.com/t/release-ikea-zigbee-drivers/123853" target="_blank">community page</a><br>