+++ ikea-zigbee-drivers/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.4.0] - 2025-01-06
+### Added
+- Add "Toggle state" option for the Power On Behavior preference - `@SplatMan`
+- LYWSD03MMC: Add support for pvvx/ZigbeeTLc firmware - `@swinokur`
+
 ## [5.3.0] - 2024-12-28
 ### Added
 - Add driver for Gewiss 2-channel Contact Interface 230V (GWA1502)